LEPv2: Refactor Kafka producer to use Long as key and add LibraryEvent producer implementation

+++ library-events-producer-v2/src/main/java/com/arshad/config/KafkaProducerConfig.java
@@ -3,6 +3,7 @@ package com.arshad.config;
 import com.arshad.model.LibraryEvent;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.IntegerSerializer;
+import org.apache.kafka.common.serialization.LongSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -33,10 +34,10 @@ public class KafkaProducerConfig {
     private Integer lingerMs;
 
     @Bean
-    public ProducerFactory<Integer, LibraryEvent> producerFactory() {
+    public ProducerFactory<Long, LibraryEvent> producerFactory() {
         Map<String, Object> configProps = new HashMap<>();
         configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, IntegerSerializer.class);
+        configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
         configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
         configProps.put(ProducerConfig.ACKS_CONFIG, acks);
         configProps.put(ProducerConfig.RETRIES_CONFIG, retries);
@@ -47,7 +48,7 @@ public class KafkaProducerConfig {
     }
 
     @Bean
-    public KafkaTemplate<Integer, LibraryEvent> kafkaTemplate() {
+    public KafkaTemplate<Long, LibraryEvent> kafkaTemplate() {
         return new KafkaTemplate<>(producerFactory());
     }
 }
